+++ packages/components/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Bug Fixes
+
+-   `Modal`: Fix modal headings and labels to use proper editor text color instead of wp-admin colors ([#71311](https://github.com/WordPress/gutenberg/pull/71311)).
+
+### Internal
+
+-   `ValidatedCheckboxControl`: Expose the component under private API's ([#71505](https://github.com/WordPress/gutenberg/pull/71505/)).
+
 ## 30.3.0 (2025-09-03)
 
 ### Bug Fixes
@@ -13,7 +21,6 @@
 
 -   Upgrade `framer-motion` package to version `^11.15.0` ([#71442](https://github.com/WordPress/gutenberg/pull/71442)).
 -   Display names for Context providers [#71208](https://github.com/WordPress/gutenberg/pull/71208).
-
 
 ## 30.2.0 (2025-08-20)